Split pure functions in reducer to helpers file
Also added reducer tests to 100%
